requirements.txt
apache-airflow-providers-http==4.6.0
apache-airflow-providers-postgres==5.11.0
psycopg2-binary>=2.9